Refactors ProcessAndSendTask to use FormSubmissionHelper
