+++ app/src/org/commcare/sync/CancellationChecker.java
@@ -1,0 +1,9 @@
+package org.commcare.sync;
+
+public interface CancellationChecker {
+
+    /**
+     * @return if the process implementing the CancellationChecker was cancelled
+     */
+    boolean wasProcessCancelled();
+}
