+++ app/src/org/commcare/sync/FormSubmissionProgressListener.java
@@ -1,0 +1,5 @@
+package org.commcare.sync;
+
+public interface FormSubmissionProgressListener {
+    void publishUpdateProgress(Long... progress);
+}
