+++ app/src/org/commcare/sync/ProcessAndSendTask.java
@@ -43,13 +43,20 @@ import java.util.Queue;
 
 import javax.crypto.spec.SecretKeySpec;
 
+import static org.commcare.sync.FormSubmissionHelper.PROGRESS_ALL_PROCESSED;
+import static org.commcare.sync.FormSubmissionHelper.SUBMISSION_BEGIN;
+import static org.commcare.sync.FormSubmissionHelper.SUBMISSION_DONE;
+import static org.commcare.sync.FormSubmissionHelper.SUBMISSION_NOTIFY;
+import static org.commcare.sync.FormSubmissionHelper.SUBMISSION_START;
+import static org.commcare.sync.FormSubmissionHelper.SUBMISSION_SUCCESS;
+
 /**
  * @author ctsims
  */
-public abstract class ProcessAndSendTask<R> extends CommCareTask<FormRecord, Long, FormUploadResult, R> implements DataSubmissionListener {
+public abstract class ProcessAndSendTask<R> extends CommCareTask<FormRecord, Long, FormUploadResult, R>
+        implements CancellationChecker, FormSubmissionProgressListener {
 
-    private String url;
-    private FormUploadResult[] results;
+    private final FormSubmissionHelper mFormSubmissionHelper;
 
     private final int sendTaskId;
 
@@ -58,25 +65,12 @@ public abstract class ProcessAndSendTask<R> extends CommCareTask<FormRecord, Lon
     public static final int PROCESSING_PHASE_ID_NO_DIALOG = -8;
     public static final int SEND_PHASE_ID_NO_DIALOG = -9;
 
-    public static final long PROGRESS_ALL_PROCESSED = 8;
-
-    public static final long SUBMISSION_BEGIN = 16;
-    public static final long SUBMISSION_START = 32;
-    public static final long SUBMISSION_NOTIFY = 64;
-    public static final long SUBMISSION_DONE = 128;
-
-    private static final long SUBMISSION_SUCCESS = 1;
-    private static final long SUBMISSION_FAIL = 0;
-
     private FormSubmissionProgressBarListener progressBarListener;
-    private List<DataSubmissionListener> formSubmissionListeners;
-    private final FormRecordProcessor processor;
 
-    private static final int SUBMISSION_ATTEMPTS = 2;
+    private List<DataSubmissionListener> formSubmissionListeners = new ArrayList<>();
 
-    private static final Queue<ProcessAndSendTask> processTasks = new LinkedList<>();
 
-    public ProcessAndSendTask(Context c, String url) {
+    protected ProcessAndSendTask(Context c, String url) {
         this(c, url, true);
     }
 
@@ -84,9 +78,8 @@ public abstract class ProcessAndSendTask<R> extends CommCareTask<FormRecord, Lon
      * @param inSyncMode blocks the user with a sync dialog
      */
     public ProcessAndSendTask(Context c, String url, boolean inSyncMode) {
-        this.url = url;
-        this.processor = new FormRecordProcessor(c);
-        this.formSubmissionListeners = new ArrayList<>();
+        mFormSubmissionHelper = new FormSubmissionHelper(c, url, this, this);
+
         if (inSyncMode) {
             this.sendTaskId = SEND_PHASE_ID;
             this.taskId = PROCESSING_PHASE_ID;
@@ -98,372 +91,7 @@ public abstract class ProcessAndSendTask<R> extends CommCareTask<FormRecord, Lon
 
     @Override
     protected FormUploadResult doTaskBackground(FormRecord... records) {
-        boolean wroteErrorToLogs = false;
-        try {
-            results = new FormUploadResult[records.length];
-            for (int i = 0; i < records.length; ++i) {
-                //Assume failure
-                results[i] = FormUploadResult.FAILURE;
-            }
-            //The first thing we need to do is make sure everything is processed,
-            //we can't actually proceed before that.
-            try {
-                wroteErrorToLogs = checkFormRecordStatus(records);
-            } catch (FileNotFoundException e) {
-                return FormUploadResult.PROGRESS_SDCARD_REMOVED;
-            } catch (TaskCancelledException e) {
-                return FormUploadResult.FAILURE;
-            }
-
-
-            this.publishProgress(PROGRESS_ALL_PROCESSED);
-
-            //Put us on the queue!
-            synchronized (processTasks) {
-                processTasks.add(this);
-            }
-            boolean needToRefresh;
-            try {
-                needToRefresh = blockUntilTopOfQueue();
-            } catch (TaskCancelledException e) {
-                return FormUploadResult.FAILURE;
-            }
-
-
-            if (needToRefresh) {
-                //There was another activity before this one. Refresh our models in case
-                //they were updated
-                for (int i = 0; i < records.length; ++i) {
-                    int dbId = records[i].getID();
-                    records[i] = processor.getRecord(dbId);
-                }
-            }
-
-            // Ok, all forms are now processed. Time to focus on sending
-            dispatchBeginSubmissionProcessToListeners(records.length);
-
-            try {
-                sendForms(records);
-            } catch (TaskCancelledException e) {
-                return FormUploadResult.FAILURE;
-            }
-
-            return FormUploadResult.getWorstResult(results);
-        } catch (SessionUnavailableException sue) {
-            this.cancel(false);
-            return FormUploadResult.PROGRESS_LOGGED_OUT;
-        } finally {
-            boolean success =
-                    FormUploadResult.FULL_SUCCESS.equals(FormUploadResult.getWorstResult(results));
-            this.endSubmissionProcess(success);
-
-            synchronized (processTasks) {
-                processTasks.remove(this);
-            }
-
-            if (success || wroteErrorToLogs) {
-                // Try to send logs if we either know we have a good connection, or know we wrote
-                // an error to the logs during form submission attempt
-                CommCareApplication.instance().notifyLogsPending();
-            }
-        }
-    }
-
-    private boolean checkFormRecordStatus(FormRecord[] records)
-            throws FileNotFoundException, TaskCancelledException {
-        boolean wroteErrorToLogs = false;
-        processor.beginBulkSubmit();
-        for (int i = 0; i < records.length; ++i) {
-            if (isCancelled()) {
-                throw new TaskCancelledException();
-            }
-            FormRecord record = records[i];
-
-            //If the form is complete, but unprocessed, process it.
-            if (FormRecord.STATUS_COMPLETE.equals(record.getStatus())) {
-                SQLiteDatabase userDb =
-                        CommCareApplication.instance().getUserDbHandle();
-                try {
-                    userDb.beginTransaction();
-                    try {
-                        records[i] = processor.process(record);
-                        userDb.setTransactionSuccessful();
-                    } finally {
-                        userDb.endTransaction();
-                    }
-                } catch (InvalidStructureException | XmlPullParserException |
-                        UnfullfilledRequirementsException e) {
-                    records[i] = handleExceptionFromFormProcessing(record, e);
-                    wroteErrorToLogs = true;
-                } catch (FileNotFoundException e) {
-                    if (CommCareApplication.instance().isStorageAvailable()) {
-                        //If storage is available generally, this is a bug in the app design
-                        Logger.log(LogTypes.TYPE_ERROR_DESIGN,
-                                "Removing form record because file was missing|" + getExceptionText(e));
-                        record.logPendingDeletion(TAG,
-                                "the xml submission file associated with the record could not be found");
-                        FormRecordCleanupTask.wipeRecord(record);
-                        records[i] = FormRecord.StandInForDeletedRecord();
-                        wroteErrorToLogs = true;
-                    } else {
-                        CommCareApplication.notificationManager().reportNotificationMessage(
-                                NotificationMessageFactory.message(ProcessIssues.StorageRemoved), true);
-                        //Otherwise, the SD card just got removed, and we need to bail anyway.
-                        throw e;
-                    }
-                } catch (IOException e) {
-                    Logger.log(LogTypes.TYPE_ERROR_WORKFLOW, "IO Issues processing a form. " +
-                            "Tentatively not removing in case they are resolvable|" + getExceptionText(e));
-                    wroteErrorToLogs = true;
-                }
-            }
-        }
-        processor.closeBulkSubmit();
-        return wroteErrorToLogs;
-    }
-
-    private FormRecord handleExceptionFromFormProcessing(FormRecord record, Exception e) {
-        String logMessage = "";
-        if (e instanceof InvalidStructureException) {
-            logMessage =
-                    String.format("Quarantining form record with ID %s due to transaction data|",
-                            record.getInstanceID());
-        } else if (e instanceof XmlPullParserException) {
-            logMessage =
-                    String.format("Quarantining form record with ID %s due to bad xml|",
-                            record.getInstanceID());
-        } else if (e instanceof UnfullfilledRequirementsException) {
-            logMessage =
-                    String.format("Quarantining form record with ID %s due to bad requirements|",
-                            record.getInstanceID());
-        }
-        logMessage = logMessage + getExceptionText(e);
-
-        CommCareApplication.notificationManager().reportNotificationMessage(
-                NotificationMessageFactory.message(ProcessIssues.BadTransactions), true);
-        Logger.log(LogTypes.TYPE_ERROR_DESIGN, logMessage);
-
-        return quarantineRecord(record, FormRecord.QuarantineReason_LOCAL_PROCESSING_ERROR);
-    }
-
-    private boolean blockUntilTopOfQueue() throws TaskCancelledException {
-        boolean needToRefresh = false;
-        while (true) {
-            //See if it's our turn to go
-            synchronized (processTasks) {
-                if (isCancelled()) {
-                    processTasks.remove(this);
-                    throw new TaskCancelledException();
-                }
-                //Are we at the head of the queue?
-                ProcessAndSendTask head = processTasks.peek();
-                if (head == this) {
-                    break;
-                }
-                //Otherwise, is the head of the queue busted?
-                //*sigh*. Apparently Cancelled doesn't result in the task status being set
-                //to !Running for reasons which baffle me.
-                if (head.getStatus() != AsyncTask.Status.RUNNING || head.isCancelled()) {
-                    //If so, get rid of it
-                    processTasks.poll();
-                }
-            }
-            //If it's not yet quite our turn, take a nap
-            try {
-                needToRefresh = true;
-                Thread.sleep(500);
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-        }
-        return needToRefresh;
-    }
-
-    private void sendForms(FormRecord[] records) throws TaskCancelledException {
-        for (int i = 0; i < records.length; ++i) {
-
-            if (previousFailurePredictsFutureFailures(results, i)) {
-                Logger.log(LogTypes.TYPE_WARNING_NETWORK,
-                        "Cancelling submission due to network errors. " + (i - 1) + " forms successfully sent.");
-                break;
-            }
-
-            if (isCancelled()) {
-                Logger.log(LogTypes.TYPE_USER, "Cancelling submission due to a manual stop. " + (i - 1) + " forms succesfully sent.");
-                throw new TaskCancelledException();
-            }
-
-            FormRecord record = records[i];
-            try {
-                if (FormRecord.STATUS_UNSENT.equals(record.getStatus())) {
-                    File folder;
-
-                    //Good!
-                    //Time to Send!
-                    try {
-                        try {
-                            if (StringUtils.isEmpty(record.getFilePath())) {
-                                throw new FileNotFoundException("File path empty for formrecord " +
-                                        record.getID() + " with xmlns " + record.getFormNamespace());
-                            }
-                            folder = new File(record.getFilePath()).getCanonicalFile().getParentFile();
-                        } catch (FileNotFoundException e) {
-                            //This will put us in the same "Missing Form" handling path as below
-                            throw e;
-                        } catch (IOException e) {
-                            // Unexpected/Unknown IO Error path from cannonical file
-                            Logger.log(LogTypes.TYPE_ERROR_WORKFLOW, "Bizarre. Exception just getting the file reference. Not removing." + getExceptionText(e));
-                            continue;
-                        }
-
-                        User user = CommCareApplication.instance().getSession().getLoggedInUser();
-                        int attemptsMade = 0;
-                        logSubmissionAttempt(record);
-                        while (attemptsMade < SUBMISSION_ATTEMPTS) {
-
-                            if (isCancelled()) {
-                                Logger.log(LogTypes.TYPE_USER, "Cancelling submission due to a manual stop. " + (i - 1) + " forms succesfully sent.");
-                                throw new TaskCancelledException();
-                            }
-
-                            results[i] = FormUploadUtil.sendInstance(i, folder,
-                                    new SecretKeySpec(record.getAesKey(), "AES"), url, this, user);
-                            if (results[i] == FormUploadResult.FULL_SUCCESS) {
-                                logSubmissionSuccess(record);
-                                break;
-                            } else if (results[i] == FormUploadResult.PROCESSING_FAILURE) {
-                                // A processing failure indicates that there there is no point in
-                                // trying that submission again immediately
-                                break;
-                            } else if (results[i] == FormUploadResult.RATE_LIMITED) {
-                                // Don't keep retrying, the server is rate limiting submissions
-                                break;
-                            } else {
-                                attemptsMade++;
-                            }
-                        }
-                        if (results[i] == FormUploadResult.RECORD_FAILURE ||
-                                results[i] == FormUploadResult.PROCESSING_FAILURE) {
-                            quarantineRecord(record, results[i]);
-                        }
-                    } catch (FileNotFoundException e) {
-                        if (CommCareApplication.instance().isStorageAvailable()) {
-                            // If storage is available generally, this is a bug in the app design
-                            // Log with multiple tags so we can track more easily
-                            Logger.log(LogTypes.SOFT_ASSERT, String.format(
-                                    "Removed form record with id %s because file was missing| %s",
-                                    record.getInstanceID(), getExceptionText(e)));
-                            Logger.log(LogTypes.TYPE_FORM_SUBMISSION, String.format(
-                                    "Removed form record with id %s because file was missing| %s",
-                                    record.getInstanceID(), getExceptionText(e)));
-                            record.logPendingDeletion(TAG,
-                                    "the xml submission file associated with the record was missing");
-                            quarantineRecord(record,
-                                    FormRecord.QuarantineReason_FILE_NOT_FOUND);
-                            results[i] = FormUploadResult.RECORD_FAILURE;
-                        } else {
-                            // Otherwise, the SD card just got removed, and we need to bail anyway.
-                            CommCareApplication.notificationManager().reportNotificationMessage(
-                                    NotificationMessageFactory.message(ProcessIssues.StorageRemoved), true);
-                            break;
-                        }
-                        continue;
-                    }
-
-                    Profile p = CommCareApplication.instance().getCommCarePlatform().getCurrentProfile();
-                    // Check for success
-                    if (results[i] == FormUploadResult.FULL_SUCCESS) {
-                        // Only delete if this device isn't set up to review.
-                        if (p == null || !p.isFeatureActive(Profile.FEATURE_REVIEW)) {
-                            FormRecordCleanupTask.wipeRecord(record);
-                        } else {
-                            // Otherwise save and move appropriately
-                            processor.updateRecordStatus(record, FormRecord.STATUS_SAVED);
-                        }
-                    }
-                } else if (FormRecord.STATUS_QUARANTINED.equals(record.getStatus()) ||
-                        FormRecord.STATUS_JUST_DELETED.equals(record.getStatus())) {
-                    // This record was either quarantined or deleted due to an error during the
-                    // pre-processing phase
-                    results[i] = FormUploadResult.RECORD_FAILURE;
-                } else {
-                    results[i] = FormUploadResult.FULL_SUCCESS;
-                }
-            } catch (SessionUnavailableException sue) {
-                throw sue;
-            } catch (Exception e) {
-                //Just try to skip for now. Hopefully this doesn't wreck the model :/
-                Logger.exception("Totally Unexpected Error during form submission: " + getExceptionText(e), e);
-            }
-        }
-    }
-
-    /**
-     * @param results      the array of submission results
-     * @param currentIndex - the index of the submission we are about to attempt
-     * @return true if there was a failure in submitting the previous form that indicates future
-     * submission attempts will also fail. (We permit proceeding if there was a local problem with
-     * a specific record, or a processing error with a specific record, since that is unrelated to
-     * how future submissions will fair).
-     */
-    private boolean previousFailurePredictsFutureFailures(FormUploadResult[] results, int currentIndex) {
-        if (currentIndex > 0) {
-            FormUploadResult lastResult = results[currentIndex - 1];
-            return !(lastResult == FormUploadResult.FULL_SUCCESS ||
-                    lastResult == FormUploadResult.RECORD_FAILURE ||
-                    lastResult == FormUploadResult.PROCESSING_FAILURE);
-        }
-        return false;
-    }
-
-    private FormRecord quarantineRecord(FormRecord record, FormUploadResult uploadResult) {
-        String reasonType =
-                (uploadResult == FormUploadResult.RECORD_FAILURE) ?
-                        FormRecord.QuarantineReason_RECORD_ERROR :
-                        FormRecord.QuarantineReason_SERVER_PROCESSING_ERROR;
-        record = processor.quarantineRecord(record, reasonType, uploadResult.getErrorMessage());
-        logAndNotifyQuarantine(record);
-        return record;
-    }
-
-    private FormRecord quarantineRecord(FormRecord record, String quarantineReasonType) {
-        record = processor.quarantineRecord(record, quarantineReasonType);
-        logAndNotifyQuarantine(record);
-        return record;
-    }
-
-    private static void logAndNotifyQuarantine(FormRecord record) {
-        Logger.log(LogTypes.TYPE_ERROR_STORAGE,
-                String.format("Quarantining Form Record with id %s because: %s",
-                        record.getInstanceID(),
-                        QuarantineUtil.getQuarantineReasonDisplayString(record, true)));
-
-        NotificationMessage m = QuarantineUtil.getQuarantineNotificationMessage(record);
-        if (m != null) {
-            CommCareApplication.notificationManager().reportNotificationMessage(m, true);
-        }
-    }
-
-    private static void logSubmissionAttempt(FormRecord record) {
-        String attemptMesssage = String.format(
-                "Attempting to submit form with id %1$s and submission ordering number %2$s",
-                record.getInstanceID(),
-                record.getSubmissionOrderingNumber());
-        Logger.log(LogTypes.TYPE_FORM_SUBMISSION, attemptMesssage);
-    }
-
-    private static void logSubmissionSuccess(FormRecord record) {
-        String successMessage = String.format(
-                "Successfully submitted form with id %1$s and submission ordering number %2$s",
-                record.getInstanceID(),
-                record.getSubmissionOrderingNumber());
-        Logger.log(LogTypes.TYPE_FORM_SUBMISSION, successMessage);
-    }
-
-    public static int pending() {
-        synchronized (processTasks) {
-            return processTasks.size();
-        }
+        return mFormSubmissionHelper.uploadForms(records);
     }
 
     @Override
@@ -491,15 +119,6 @@ public abstract class ProcessAndSendTask<R> extends CommCareTask<FormRecord, Lon
         }
     }
 
-    public void addProgressBarSubmissionListener(FormSubmissionProgressBarListener listener) {
-        this.progressBarListener = listener;
-        addSubmissionListener(listener);
-    }
-
-    public void addSubmissionListener(DataSubmissionListener submissionListener) {
-        formSubmissionListeners.add(submissionListener);
-    }
-
     private void dispatchBeginSubmissionProcessToListeners(int totalItems) {
         for (DataSubmissionListener listener : formSubmissionListeners) {
             listener.beginSubmissionProcess(totalItems);
@@ -524,32 +143,34 @@ public abstract class ProcessAndSendTask<R> extends CommCareTask<FormRecord, Lon
         }
     }
 
+    public void addProgressBarSubmissionListener(FormSubmissionProgressBarListener listener) {
+        this.progressBarListener = listener;
+        addSubmissionListener(listener);
+    }
+
+    public void addSubmissionListener(DataSubmissionListener submissionListener) {
+        formSubmissionListeners.add(submissionListener);
+    }
+
+    @Override
+    public boolean wasProcessCancelled() {
+//        Apparently Cancelled doesn't result in the task status being set to !Running for reasons which baffle me.
+        return getStatus() != Status.RUNNING || isCancelled();
+    }
+
     @Override
     protected void onPostExecute(FormUploadResult result) {
         super.onPostExecute(result);
-
         clearState();
     }
 
     private void clearState() {
-        url = null;
-        results = null;
+        mFormSubmissionHelper.cleanUp();
     }
 
-    protected int getSuccessfulSends() {
-        int successes = 0;
-        if (results != null) {
-            for (FormUploadResult formResult : results) {
-                if (formResult != null && FormUploadResult.FULL_SUCCESS == formResult) {
-                    successes++;
-                }
-            }
-        }
-        return successes;
-    }
 
     protected String getLabelForFormsSent() {
-        int successfulSends = getSuccessfulSends();
+        int successfulSends = mFormSubmissionHelper.getSuccessfulSends();
         String label;
         switch (successfulSends) {
             case 0:
@@ -565,41 +186,6 @@ public abstract class ProcessAndSendTask<R> extends CommCareTask<FormRecord, Lon
         return label;
     }
 
-
-    //Wrappers for the internal stuff
-    @Override
-    public void beginSubmissionProcess(int totalItems) {
-        this.publishProgress(SUBMISSION_BEGIN, (long)totalItems);
-    }
-
-    @Override
-    public void startSubmission(int itemNumber, long sizeOfItem) {
-        this.publishProgress(SUBMISSION_START, (long)itemNumber, sizeOfItem);
-    }
-
-    @Override
-    public void notifyProgress(int itemNumber, long progress) {
-        this.publishProgress(SUBMISSION_NOTIFY, (long)itemNumber, progress);
-    }
-
-    @Override
-    public void endSubmissionProcess(boolean success) {
-        if (success) {
-            this.publishProgress(SUBMISSION_DONE, SUBMISSION_SUCCESS);
-        } else {
-            this.publishProgress(SUBMISSION_DONE, SUBMISSION_FAIL);
-        }
-    }
-
-    private String getExceptionText(Exception e) {
-        try {
-            ByteArrayOutputStream bos = new ByteArrayOutputStream();
-            e.printStackTrace(new PrintStream(bos));
-            return new String(bos.toByteArray());
-        } catch (Exception ex) {
-            return null;
-        }
-    }
 
     @Override
     protected void onCancelled() {
@@ -618,6 +204,11 @@ public abstract class ProcessAndSendTask<R> extends CommCareTask<FormRecord, Lon
     }
 
     @Override
+    public void publishUpdateProgress(Long... progress) {
+        publishProgress(progress);
+    }
+
+    @Override
     public void connect(CommCareTaskConnector<R> connector) {
         super.connect(connector);
         if (progressBarListener != null) {
@@ -626,6 +217,7 @@ public abstract class ProcessAndSendTask<R> extends CommCareTask<FormRecord, Lon
         }
     }
 
-    private static class TaskCancelledException extends Exception {
+    protected int getSuccessfulSends() {
+        return mFormSubmissionHelper.getSuccessfulSends();
     }
 }
